+++ src/main/java/org/scion/jpan/ScionService.java
@@ -307,10 +307,8 @@ public class ScionService {
    * @return All paths returned by the path service.
    * @throws IOException if an errors occurs while querying paths.
    */
-  @Deprecated // Please use lookUp() instead
+  @Deprecated // Please use lookup() instead
   public List<RequestPath> getPaths(InetSocketAddress dstAddress) throws IOException {
-    // TODO this is ambiguous .... InetSocketAddress argument
-    // Use getHostString() to avoid DNS reverse lookup.
     ScionAddress sa = getScionAddress(dstAddress.getHostString());
     return getPaths(sa.getIsdAs(), sa.getInetAddress(), dstAddress.getPort());
   }
@@ -322,15 +320,7 @@ public class ScionService {
    * @param dstScionAddress Destination IP address. Must belong to a SCION enabled end host.
    * @return All paths returned by the path service.
    */
-  // TODO this is bad!!! Why ISD + SCION-Address?????
-  // TODO this is bad!!! Why ISD + SCION-Address?????
-  // TODO this is bad!!! Why ISD + SCION-Address?????
-  // TODO this is bad!!! Why ISD + SCION-Address?????
-  // TODO this is bad!!! Why ISD + SCION-Address?????
-  // TODO this is bad!!! Why ISD + SCION-Address?????
-  // TODO this is bad!!! Why ISD + SCION-Address?????
   public List<RequestPath> getPaths(long dstIsdAs, InetSocketAddress dstScionAddress) {
-    // TODO change method API name to make clear that this requires a SCION IP.
     return getPaths(dstIsdAs, dstScionAddress.getAddress(), dstScionAddress.getPort());
   }
 
@@ -341,19 +331,35 @@ public class ScionService {
    * @return All paths returned by the path service.
    */
   public List<RequestPath> getPaths(RequestPath path) {
-    return getPaths(path.getIsdAs(), path.getAddress(), path.getPort());
+    return getPaths(path.getIsdAs(), path.getRemoteAddress(), path.getRemotePort());
   }
 
   /**
    * Request paths from the local ISD/AS to the destination.
    *
    * @param dstIsdAs Destination ISD/AS
-   * @param dstAddress Destination IP address
+   * @param dstAddress A SCION-enabled Destination IP address
    * @param dstPort Destination port
    * @return All paths returned by the path service. Returns an empty list if no paths are found.
    */
   public List<RequestPath> getPaths(long dstIsdAs, InetAddress dstAddress, int dstPort) {
-    return getPaths(RequestPath.fromScionIP(dstIsdAs, dstAddress, dstPort));
+    return getPaths(ScionAddress.create(dstIsdAs, dstAddress), dstPort);
+  }
+
+  /**
+   * Resolves the address to a SCION address, request paths, and selects a path using the policy.
+   *
+   * @param dstAddr Destination address
+   * @param policy path policy
+   * @return All paths returned by the path service.
+   * @throws ScionException if the DNS/TXT lookup did not return a (valid) SCION address.
+   */
+  public RequestPath lookupAndGetPath(InetSocketAddress dstAddr, PathPolicy policy)
+      throws ScionException {
+    if (policy == null) {
+      policy = PathPolicy.DEFAULT;
+    }
+    return policy.filter(getPaths(lookupAddress(dstAddr.getHostString()), dstAddr.getPort()));
   }
 
   /**
@@ -362,12 +368,14 @@ public class ScionService {
    * @param dstAddress Destination SCION address
    * @return All paths returned by the path service.
    */
-  public List<RequestPath> getPaths(Path dstAddress) {
+  public List<RequestPath> getPaths(ScionAddress dstAddress, int dstPort) {
     long srcIsdAs = getLocalIsdAs();
     List<Daemon.Path> paths = getPathList(srcIsdAs, dstAddress.getIsdAs());
     List<RequestPath> scionPaths = new ArrayList<>(paths.size());
     for (int i = 0; i < paths.size(); i++) {
-      scionPaths.add(RequestPath.create(paths.get(i), dstAddress));
+      scionPaths.add(
+          RequestPath.create(
+              paths.get(i), dstAddress.getIsdAs(), dstAddress.getInetAddress(), dstPort));
     }
     return scionPaths;
   }
@@ -439,25 +447,14 @@ public class ScionService {
   }
 
   /**
-   * Uses DNS and hostfiles to look up a SCIOn enabled IP address for a give host string.
-   *
-   * @param hostName hostName of the host to resolve
-   * @return A ScionSocketAddress
-   * @throws ScionException if the DNS/TXT lookup did not return a (valid) SCION address.
-   */
-  public Path lookupSocketAddress(String hostName, int port) throws ScionException {
-    return RequestPath.fromScionAddress(lookupAddress(hostName), port);
-  }
-
-  /**
-   * Uses DNS and hostfiles to look up a SCIOn enabled IP address for a give host string.
+   * Uses DNS and hostfiles to look up a SCION enabled IP address for a give host string.
    *
    * @param hostName hostName of the host to resolve
    * @return A ScionAddress
    * @throws ScionException if the DNS/TXT lookup did not return a (valid) SCION address.
    */
   @Deprecated // TODO NOW: remove this lookupAddress() method in favor of lookupSocketAddress()
-  private ScionAddress lookupAddress(String hostName) throws ScionException {
+  ScionAddress lookupAddress(String hostName) throws ScionException {
     ScionAddress scionAddress = scionAddressCache.get(hostName);
     if (scionAddress != null) {
       return scionAddress;
@@ -605,7 +602,7 @@ public class ScionService {
     }
     // expired, get new path
     RequestPath newPath = pathPolicy.filter(getPaths(path));
-    path.setDetails(newPath.getDetails());
+    ((RequestPathImpl) path).setDetails(newPath.getDetails());
     return true;
   }
 }